Refactor core orchestration into modular package
