+++ src/datacore/quality/__init__.py
@@ -1,6 +1,6 @@
 """Lightweight data-quality helpers."""
 
-from .expectations import (
+from mvp_config_driven.core.quality import (  # noqa: F401
     DQReport,
     ExpectationResult,
     apply_expectations,
