+++ src/mvp_config_driven/__init__.py
@@ -1,0 +1,13 @@
+"""Core configuration driven orchestration primitives."""
+
+from .core.config import LayerConfig
+from .core.context import PipelineContext, build_context, context_paths, ensure_engine, stop_engine
+
+__all__ = [
+    "LayerConfig",
+    "PipelineContext",
+    "build_context",
+    "context_paths",
+    "ensure_engine",
+    "stop_engine",
+]
