+++ src/mvp_config_driven/adapters/__init__.py
@@ -1,0 +1,3 @@
+"""Adapter packages bridging ports with concrete implementations."""
+
+__all__ = []
