+++ src/mvp_config_driven/core/quality/__init__.py
@@ -1,0 +1,15 @@
+"""Core data-quality helpers."""
+
+from .expectations import (
+    DQReport,
+    ExpectationResult,
+    apply_expectations,
+    evaluate_expectations,
+)
+
+__all__ = [
+    "DQReport",
+    "ExpectationResult",
+    "apply_expectations",
+    "evaluate_expectations",
+]
