+++ src/mvp_config_driven/ports/__init__.py
@@ -1,0 +1,15 @@
+"""Public interfaces required by orchestration adapters."""
+
+from .jobs import JobBackendPort
+from .secrets import SecretsPort
+from .storage import StoragePort
+from .streaming import StreamingPort
+from .tables import TablesPort
+
+__all__ = [
+    "JobBackendPort",
+    "SecretsPort",
+    "StoragePort",
+    "StreamingPort",
+    "TablesPort",
+]
